extract service handler to own file
